refactor(gateway): clean code, move file structure (#1371)
* refactor(gateway): clean code, move file structure

* refactor(proto): jina primitive types
